pyproject.toml
[tool.poetry]
name = "glider"
version = "0.1.0"
description = "Generalist image Detectors"
authors = ["knowledgator.com"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
torch = "^2.0.0"
transformers = "^4.37.2"
numpy = "^1.26.4"
pillow = "^10.0.0" 

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"